requirements.txt
aiogram>=3.0.0,<3.7
SQLAlchemy>=1.4
aiosqlite>=0.17
python-dotenv>=1.0